requirements.txt
# Core dependencies
langchain==0.1.0
langchain-google-genai==0.0.5
langchain-community==0.0.10
streamlit==1.30.0
python-dotenv==1.0.0
tenacity==8.2.3
tavily-python==0.2.8
google-generativeai==0.3.1
pandas==2.1.0

# Additional dependencies for Streamlit Cloud
numpy==1.24.3
protobuf==4.25.1
typing-extensions==4.8.0
pydantic==2.5.2
requests==2.31.0
aiohttp==3.9.1
python-dateutil==2.8.2
pytz==2023.3
tqdm==4.66.1
packaging==23.2
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
urllib3==2.1.0
click==8.1.7
watchdog==3.0.0
gitpython==3.1.40
validators==0.22.0
beautifulsoup4==4.12.2
soupsieve==2.5
lxml==4.9.3
html5lib==1.1 